Drop debian/patches/0007-udev-seat-tag.patch and ensure seat tag is properly applied instead
Drop debian/patches/0007-udev-seat-tag.patch and add dependency to udev
and systemd (>= 232-8~). Starting with these version, the 71-seat.rules
file is installed in the initramfs. This makes sure that the seat tag
is set on the needed devices.

This doesn't require the user to run systemd as PID1, juste merely
having the systemd packag installed.